refactor: rename CloudflareKVInterface to CloudflareKVStore
Renamed to better reflect its role as a concrete implementation of IKVStore rather than an interface definition. Updated related imports and documentation.
